+++ src/http/http.js
@@ -10,6 +10,8 @@ var regex = require( './regex.js' );
 var Router = express.Router;
 var expreq = express.request;
 var expres = express.response;
+expressInit.name = 'expressInit';
+queryParser.name = 'queryParser';
 
 function buildUrl() {
 	var idx = 0,
@@ -171,9 +173,9 @@ function queryParser( req, res, next ) {
 	next();
 }
 
-function registerMiddleware( state, filter, callback ) {
+function registerMiddleware( state, filter, callback, alias ) {
 	var fn = function( router ) {
-		log.debug( 'MIDDLEWARE: %s mounted at %s', ( callback.name || 'anonymous' ), filter );
+		log.debug( 'SYSTEM MIDDLEWARE: %s mounted at %s', ( callback.name || alias || 'anonymous' ), filter );
 		router.use( filter, callback );
 	};
 	if ( state.app ) {
@@ -182,9 +184,11 @@ function registerMiddleware( state, filter, callback ) {
 	state.systemMiddleware.push( fn );
 }
 
-function registerUserMiddleware( state, filter, callback ) {
+function registerUserMiddleware( state, filter, callback, alias ) {
+	// if using an auth strategy, move cookie and session middleware before passport middleware
+	// to take advantage of sessions/cookies and avoid authenticating on every request
 	var fn = function( router ) {
-		log.debug( 'MIDDLEWARE: %s mounted at %s', ( callback.name || 'anonymous' ), filter );
+		log.debug( 'USER MIDDLEWARE: %s mounted at %s', ( callback.name || alias || 'anonymous' ), filter );
 		router.use( filter, callback );
 	};
 	if ( state.app ) {
@@ -194,7 +198,7 @@ function registerUserMiddleware( state, filter, callback ) {
 }
 
 function registerRoute( state, url, method, callback ) {
-	method = method.toLowerCase();
+	method = method ? method.toLowerCase() : 'all';
 	method = method === 'all' || method === 'any' ? 'all' : method;
 	var fn = function() {
 		url = prefix( state, url );
@@ -238,19 +242,24 @@ function start( state, config, passport ) {
 	state.config = config;
 	state.passport = passport;
 	if ( config.parseAhead ) {
-		registerMiddleware( state, '/', preprocessPathVariables.bind( undefined, state ) );
+		registerMiddleware(
+			state,
+			'/',
+			preprocessPathVariables.bind( undefined, state ),
+			'preprocessPathVariables'
+		);
 	}
 	// if using an auth strategy, move cookie and session middleware before passport middleware
 	// to take advantage of sessions/cookies and avoid authenticating on every request
 	if ( passport ) {
-		state.middlewareLib.useCookies( state.middleware );
-		state.middlewareLib.useSession( state.middleware );
+		state.middlewareLib.useCookies( state.ahMiddleware );
+		state.middlewareLib.useSession( state.ahMiddleware );
 		_.each( passport.getMiddleware( '/' ), function( m ) {
-			state.middleware( m.path, m.fn );
+			state.ahMiddleware( m.path, m.fn, m.alias );
 		} );
 	}
 	// prime middleware with defaults
-	state.middlewareLib.attach( state.middleware, passport !== undefined );
+	state.middlewareLib.attach( state.ahMiddleware, passport !== undefined );
 
 	initializePublicRoute( state );
 	state.app = express();
@@ -296,6 +305,7 @@ module.exports = function( request, middleware ) {
 		start: start.bind( undefined, state ),
 		static: registerStaticPath.bind( undefined, state ),
 		stop: stop.bind( undefined, state ),
+		ahMiddleware: registerMiddleware.bind( undefined, state )
 	} );
 	return state;
 };