+++ src/http/middleware.js
@@ -1,7 +1,6 @@
 var _ = require( 'lodash' );
 var bodyParser = require( 'body-parser' );
 var cookies = require( 'cookie-parser' );
-var sessionLib = require( 'express-session' );
 var multer = require( 'multer' );
 var metronic = require( '../metrics' );
 var os = require( 'os' );
@@ -14,9 +13,9 @@ function applyCookieMiddleware( state, attach ) {
 	}
 }
 
-function applyMiddelware( state, attach, hasAuth ) {
+function applyMiddleware( state, attach, hasAuth ) {
 	// add a timer to track ALL requests
-	attach( '/', requestMetrics.bind( undefined, state ) );
+	attach( '/', requestMetrics.bind( undefined, state ), 'metrics' );
 
 	if ( !hasAuth ) {
 		applyCookieMiddleware( state, attach );
@@ -29,7 +28,7 @@ function applyMiddelware( state, attach, hasAuth ) {
 		attach( '/', bodyParser.json( { type: 'application/vnd.api+json' } ) );
 		attach( '/', multer( {
 			dest: state.config.tmp
-		} ) );
+		} ), 'multer' );
 	}
 
 	if ( !hasAuth ) {
@@ -66,14 +65,14 @@ function configure( state, config ) {
 		name: 'ah.sid',
 		secret: 'autohostthing',
 		resave: true,
-		store: new sessionLib.MemoryStore(),
+		store: new state.sessionLib.MemoryStore(),
 		saveUninitialized: true,
 		rolling: false
 	};
 	var cookieConfig = _.defaults( state.config.cookie || {}, cookieDefaults );
 	var sessionConfig = _.defaults( state.config.session || {}, sessionDefaults );
 	sessionConfig.cookie = cookieConfig;
-	state.session = sessionLib( sessionConfig );
+	state.session = state.sessionLib( sessionConfig );
 }
 
 function requestMetrics( state, req, res, next ) {
@@ -138,7 +137,7 @@ function requestMetrics( state, req, res, next ) {
 	next();
 }
 
-module.exports = function() {
+module.exports = function( sessionLib ) {
 	var state = {
 		config: undefined,
 		cookieParser: cookies(),
@@ -147,7 +146,7 @@ module.exports = function() {
 		sessionLib: sessionLib
 	};
 	_.merge( state, {
-		attach: applyMiddelware.bind( undefined, state ),
+		attach: applyMiddleware.bind( undefined, state ),
 		configure: configure.bind( undefined, state ),
 		useCookies: applyCookieMiddleware.bind( undefined, state ),
 		useSession: applySessionMiddleware.bind( undefined, state ),